fix music loader. add internal unassigned template. add metadata viewer

+++ src/main/java/fr/poulpogaz/musicdl/ui/MusicdlFrame.java
@@ -139,7 +139,7 @@ public class MusicdlFrame extends JFrame {
         return bottomBar;
     }
 
-    public void setLoadingMusicCount(int count) {
+    public void setLoadingMusicCount(long count) {
         loadingMusicsLabel.setText(count + " musics loading");
     }
